chore: migrate unit tests to Vitest
Replace per-file scripts/test-*.mjs runners with colocated src/**/*.test.ts,
expose pnpm test / test:watch, and point CI/docs at the standard entry.
