+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -4,7 +4,7 @@ Why this change is needed (not only what files changed).
 
 ## Checklist
 
-- [ ] `pnpm format:check && pnpm lint && pnpm test-only && pnpm typecheck && pnpm build` passes locally
+- [ ] `pnpm format:check && pnpm lint && pnpm test && pnpm typecheck && pnpm build` passes locally
 - [ ] UI / capture changes were smoke-tested (`pnpm demo` or a real page)
 - [ ] Shared logic changes include or update tests when practical
 - [ ] No secrets or personal data in fixtures / screenshots
