+++ package.json
@@ -13,24 +13,8 @@
     "dev": "node scripts/build.mjs --watch",
     "build": "tsc --noEmit && node scripts/build.mjs",
     "typecheck": "tsc --noEmit",
-    "test:parser": "node scripts/test-parser.mjs",
-    "test:connect": "node scripts/test-connect.mjs",
-    "test:export": "node scripts/test-export.mjs",
-    "test:spec": "node scripts/test-spec.mjs",
-    "test:timing": "node scripts/test-timing.mjs",
-    "test:request-view": "node scripts/test-request-view.mjs",
-    "test:close": "node scripts/test-close.mjs",
-    "test:ai-merge": "node scripts/test-ai-merge.mjs",
-    "test:events-virtual": "node scripts/test-events-virtual.mjs",
-    "test:conversation-text": "node scripts/test-conversation-text.mjs",
-    "test:conversation-virtual": "node scripts/test-conversation-virtual.mjs",
-    "test:merge-session": "node scripts/test-merge-session.mjs",
-    "test:headers": "node scripts/test-headers.mjs",
-    "test:event-stamp": "node scripts/test-event-stamp.mjs",
-    "test:relay-buffer": "node scripts/test-relay-buffer.mjs",
-    "test:detect": "node scripts/test-detect.mjs",
-    "test:eventsource-on": "node scripts/test-eventsource-on.mjs",
-    "test-only": "pnpm test:parser && pnpm test:connect && pnpm test:export && pnpm test:spec && pnpm test:timing && pnpm test:request-view && pnpm test:close && pnpm test:ai-merge && pnpm test:events-virtual && pnpm test:conversation-text && pnpm test:conversation-virtual && pnpm test:merge-session && pnpm test:headers && pnpm test:event-stamp && pnpm test:relay-buffer && pnpm test:detect && pnpm test:eventsource-on",
+    "test": "vitest run",
+    "test:watch": "vitest",
     "lint": "eslint src --max-warnings=0",
     "lint:fix": "eslint src --fix",
     "format": "prettier --write .",
@@ -47,7 +31,8 @@
     "prettier": "^3.9.6",
     "typescript": "^5.7.3",
     "typescript-eslint": "^8.66.0",
-    "vite": "^6.0.7"
+    "vite": "^6.0.7",
+    "vitest": "^3.2.7"
   },
   "pnpm": {
     "onlyBuiltDependencies": [
