+++ tsconfig.json
@@ -11,5 +11,6 @@
     "resolveJsonModule": true,
     "types": ["chrome"]
   },
-  "include": ["src/**/*.ts", "src/**/*.d.ts"]
+  "include": ["src/**/*.ts", "src/**/*.d.ts", "vitest.config.ts"],
+  "exclude": ["src/**/*.test.ts"]
 }
